+++ vnext/Microsoft.ReactNative/NativeModulesProvider.cpp
@@ -4,7 +4,6 @@
 #include "pch.h"
 #include "NativeModulesProvider.h"
 #include "ABICxxModule.h"
-#include "ABIModule.h"
 #include "IReactModuleBuilder.h"
 
 #include <ReactUWP/ReactUwp.h>
@@ -24,20 +23,12 @@ NativeModulesProvider::GetModules(
   // std::shared_ptr<facebook::react::MessageQueueThread>
   // queueThread(defaultQueueThread);
   std::vector<facebook::react::NativeModuleDescription> modules;
-  modules.reserve(m_modules.size());
 
   if (m_modulesWorkerQueue == nullptr) {
     // TODO: The queue provided is the UIMessageQueueThread which isn't needed
     // for native modules. As a workaround for now let's just use a new worker
     // message queue.
     m_modulesWorkerQueue = react::uwp::CreateWorkerMessageQueue();
-  }
-
-  for (auto &module : m_modules) {
-    modules.emplace_back(
-        winrt::to_string(module.Name()),
-        [module]() { return std::make_unique<ABIModule>(module); },
-        m_modulesWorkerQueue);
   }
 
   for (auto &entry : m_moduleProviders) {
@@ -56,15 +47,6 @@ NativeModulesProvider::GetModules(
 }
 
 NativeModulesProvider::NativeModulesProvider() noexcept {}
-
-void NativeModulesProvider::RegisterModule(INativeModule const &module) {
-  // TODO: This is taking a naive approach right now and just adding
-  // everything. Consider whether to add the CanOverrideExistingModule on
-  // INativeModule and then check it here to see whether a module being
-  // registered is allowed to take precedence over one that was already
-  // registered.
-  m_modules.push_back(module);
-}
 
 void NativeModulesProvider::AddModuleProvider(
     winrt::hstring const &moduleName,