+++ vnext/Microsoft.ReactNative/ReactInstanceManagerBuilder.cpp
@@ -29,7 +29,6 @@ ReactInstanceManagerBuilder::Build() {
       m_instanceSettings,
       to_string(m_jsBundleFile),
       to_string(m_jsMainModuleName),
-      m_packages,
       m_packageProviders,
       m_useDeveloperSupport,
       m_initialLifecycleState);