+++ vnext/ReactUWP/Views/PopupViewManager.cpp
@@ -8,6 +8,7 @@
 #include "TouchEventHandler.h"
 
 #include <Modules/NativeUIManager.h>
+#include <Utils/Helpers.h>
 #include <Utils/ValueUtils.h>
 #include <winrt/Windows.UI.Core.h>
 
@@ -73,8 +74,23 @@ void PopupShadowNode::createView() {
   m_popupOpenedRevoker =
       popup.Opened(winrt::auto_revoke, [=](auto &&, auto &&) {
         auto instance = wkinstance.lock();
-        if (!m_updating && instance != nullptr)
+        if (!m_updating && instance != nullptr) {
           SetAnchorPosition(popup);
+
+          // When multiple flyouts/popups are overlapping, XAML's theme shadows
+          // don't render properly. As a workaround we enable a z-index
+          // translation based on an elevation derived from the count of open
+          // popups/flyouts. We apply this translation on open of the popup.
+          // (Translation is only supported on RS5+, eg. IUIElement9)
+          if (auto uiElement9 = GetView().try_as<winrt::IUIElement9>()) {
+            auto numOpenPopups = CountOpenPopups();
+            if (numOpenPopups > 0) {
+              winrt::Numerics::float3 translation{
+                  0, 0, (float)16 * numOpenPopups};
+              popup.Translation(translation);
+            }
+          }
+        }
       });
 }